Use the share pointer , weak_ptr instead of the unquie_ptr
