+++ codecs/apv/enc/Qv2ApvEncoder.h
@@ -7,23 +7,13 @@
 #include <mutex>
 #include <condition_variable>
 #include <queue>
-#include <atomic>
 
 class Qv2ApvEncoder : public Qv2Component {
 public:
-    enum State {
-        STATE_UNINITIALIZED,
-        STATE_INITIALIZED,
-        STATE_RUNNING,
-        STATE_STOPPED,
-        STATE_ERROR
-    };
-
     Qv2ApvEncoder();
     ~Qv2ApvEncoder() override;
 
     // Qv2Component interface
-    void setListener(Listener* listener) override { mListener = listener; }
     bool queue(std::unique_ptr<Qv2Work> work) override;
     bool start() override;
     void stop() override;
@@ -32,7 +22,6 @@ public:
 
     // APV Specific setup
     int initialize(oapve_param_t* param);
-    State getState() const { return mState; }
 
 private:
     void processLoop();
@@ -46,9 +35,6 @@ private:
     std::mutex mMutex;
     std::condition_variable mCv;
     std::queue<std::unique_ptr<Qv2Work>> mWorkQueue;
-    
-    std::atomic<State> mState{STATE_UNINITIALIZED};
-    Listener* mListener = nullptr;
 };
 
 #endif // QV2APVENCODER_H
