+++ codecs/bases/Qv2Component.h
@@ -2,79 +2,49 @@
 #define QV2COMPONENT_H
 
 #include "Qv2Work.h"
+#include "Qv2Errors.h"
 #include <memory>
 #include <string>
+#include <atomic>
 
 /**
- * @brief Qv2Component: Abstract base class for all codec components (APV, H.264, etc.)
+ * @brief Qv2Component: Abstract base class for all codec components.
  */
-class Qv2Component {
+class Qv2Component : public std::enable_shared_from_this<Qv2Component> {
 public:
-    /**
-     * @brief Listener interface for asynchronous callbacks.
-     */
+    enum State {
+        STATE_UNINITIALIZED,
+        STATE_INITIALIZED,
+        STATE_RUNNING,
+        STATE_STOPPED,
+        STATE_ERROR
+    };
+
     class Listener {
     public:
         virtual ~Listener() = default;
-        
-        /**
-         * @brief Called when a work unit is completed.
-         * @param component The component that finished the work.
-         * @param work The completed work unit containing output data.
-         */
-        virtual void onWorkDone(Qv2Component* component, std::unique_ptr<Qv2Work> work) = 0;
-        
-        /**
-         * @brief Called when an error occurs during processing.
-         */
-        virtual void onError(Qv2Component* component, int error) = 0;
+        virtual void onWorkDone(std::weak_ptr<Qv2Component> component, std::unique_ptr<Qv2Work> work) = 0;
+        virtual void onError(std::weak_ptr<Qv2Component> component, int error) = 0;
     };
 
     virtual ~Qv2Component() = default;
 
-    /**
-     * @brief Set the name of the component.
-     */
     virtual void setName(const std::string& name) { mName = name; }
-
-    /**
-     * @brief Get the name of the component.
-     */
     virtual std::string getName() const { return mName; }
+    virtual State getState() const { return mState; }
 
-    /**
-     * @brief Sets the listener for receiving callbacks.
-     */
-    virtual void setListener(Listener* listener) = 0;
+    virtual void setListener(Listener* listener) { mListener = listener; }
 
-    /**
-     * @brief Queues a work unit for processing (non-blocking).
-     * @return true if the work was successfully queued.
-     */
     virtual bool queue(std::unique_ptr<Qv2Work> work) = 0;
-
-    /**
-     * @brief Starts the component processing thread.
-     */
     virtual bool start() = 0;
-
-    /**
-     * @brief Stops the component and joins the processing thread.
-     */
     virtual void stop() = 0;
-
-    /**
-     * @brief Flushes all pending work units in the queue.
-     */
     virtual void flush() = 0;
-
-    /**
-     * @brief Releases all resources held by the component.
-     */
     virtual void release() = 0;
 
 protected:
     std::string mName;
+    Listener* mListener = nullptr;
+    std::atomic<State> mState{STATE_UNINITIALIZED};
 };
 
 #endif // QV2COMPONENT_H
